DOWNSTREAM: <carry>: ETCD-656: Automate datadir move after quorum-restore
This PR will add the notion of cluster ID into the initial cluster discovery process.
This allows us to automatically archive a data directory when we detect the cluster identifier changing.
The cluster identifier will only change when we are running a restore operation.

The detection requires that the revision.json (created by the revision
monitor sidecar) contains the cluster id. The cluster identifier is also
stored in the local WAL, which is much more expensive to parse. We're going to
only fallback to it when we could not get the cluster id from the
revision.json for any reason. Otherwise the WAL stays untouched, no
repair operations are attempted when it is found corrupted.

Signed-off-by: Thomas Jungblut <tjungblu@redhat.com>